convert randomep to slash command, move info 'get_show_embed' into utils
